fix(per_support): fix comparison in aper_get_length()
From ITU X.691-0207 section $19 / Encoding of sequence-of
If there is no finite maximum or "ub" is greater than or equal to 64K we say that "ub" is unset.
So the comparison for range should be strict for the 65536 value (it is equal to 64K, so unset).

This was our bug for NGReset decoding